+++ examples/digitalWrite.py
@@ -7,8 +7,8 @@ mm = MachineMotion(DEFAULT_IP_ADDRESS.usb_windows)
 # Toggles the output pins on all connected IO Modules
 detectedIOModules = mm.detectIOModules()
 for IO_Name, IO_NetworkID in detectedIOModules.items():
-    writePins ={"Pin 1":0, "Pin 2":1, "Pin 3": 2, "Pin 4":3}
-    for writePin in writePins.keys():
+    writePins ={"Pin 1":1, "Pin 2":2, "Pin 3":3, "Pin 4":4}
+    for writePin in writePins:
         print(writePin + " on " + IO_Name + " is going to flash twice")
         
         mm.digitalWrite(IO_NetworkID, writePins[writePin], 1)
@@ -18,6 +18,3 @@ for IO_Name, IO_NetworkID in detectedIOModules.items():
         mm.digitalWrite(IO_NetworkID, writePins[writePin], 1)
         time.sleep(1)
         mm.digitalWrite(IO_NetworkID, writePins[writePin], 0)
-
-  
-